+++ haiku_rag_slim/haiku/rag/ingester/pollers/base.py
@@ -8,7 +8,7 @@ from haiku.rag.config import SourceConfig
 from haiku.rag.ingester.batch import BatchChange, BatchSourceSummary
 from haiku.rag.ingester.queue.models import JobOp, SyncRow
 from haiku.rag.ingester.queue.repository import JobRepo, SyncStateRepo
-from haiku.rag.ingester.sources.base import (
+from haiku.rag.sources.base import (
     Source,
     SourceEvent,
     SourceEventKind,
