+++ haiku_rag_slim/haiku/rag/sources/__init__.py
@@ -1,19 +1,19 @@
-from haiku.rag.ingester.sources.base import (
+from haiku.rag.sources.base import (
     FetchResult,
     RevisionSnapshot,
     Source,
     SourceEvent,
     SourceEventKind,
 )
-from haiku.rag.ingester.sources.filter import FileFilter
-from haiku.rag.ingester.sources.fs import FSSource
-from haiku.rag.ingester.sources.http import HTTPSource
-from haiku.rag.ingester.sources.registry import (
+from haiku.rag.sources.filter import FileFilter
+from haiku.rag.sources.fs import FSSource
+from haiku.rag.sources.http import HTTPSource
+from haiku.rag.sources.registry import (
     resolve_adhoc_fetcher,
     resolve_configured_source,
 )
-from haiku.rag.ingester.sources.s3 import S3Source
-from haiku.rag.ingester.sources.webdav import WebDAVSource
+from haiku.rag.sources.s3 import S3Source
+from haiku.rag.sources.webdav import WebDAVSource
 
 __all__ = [
     "FetchResult",
