+++ haiku_rag_slim/haiku/rag/sources/s3.py
@@ -5,14 +5,14 @@ from datetime import UTC, datetime
 from urllib.parse import urlparse
 
 from haiku.rag.client.exceptions import UnsupportedSourceError
-from haiku.rag.ingester.sources.base import (
+from haiku.rag.sources.base import (
     FetchResult,
     RevisionSnapshot,
     SourceEvent,
     SourceEventKind,
     check_file_size,
 )
-from haiku.rag.ingester.sources.filter import (
+from haiku.rag.sources.filter import (
     FileFilter,
     _default_supported_extensions,
 )
