+++ haiku_rag_slim/haiku/rag/sources/webdav.py
@@ -7,14 +7,14 @@ from xml.etree.ElementTree import Element, fromstring
 
 import httpx
 
-from haiku.rag.ingester.sources.base import (
+from haiku.rag.sources.base import (
     FetchResult,
     RevisionSnapshot,
     SourceEvent,
     SourceEventKind,
     check_file_size,
 )
-from haiku.rag.ingester.sources.filter import (
+from haiku.rag.sources.filter import (
     FileFilter,
     _default_supported_extensions,
 )
